#31: PlaceMapActivity.kt와 SearchPlaceActivity.kt 파일의 패키지 변경
