+++ feature_add_place/src/main/res/layout/activity_place_map.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PlaceMapActivity">
+    tools:context=".presentation.PlaceMapActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
